requirements.txt
# Claude API
anthropic>=0.25.0

# FastAPI
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# RAG & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-json-logger>=2.0.7

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.2

# Data processing
pandas>=2.1.0
numpy>=1.24.0
